+++ packages/terra-site/src/TestLinks.jsx
@@ -6,6 +6,7 @@ import styles from './site.scss';
 const TestLinks = () => (
   <div className={styles.app} >
     <ul className={styles['site-nav']}>
+      <li><Link to="/tests/alert-tests">Alert Tests</Link></li>
       <li><Link to="/tests/arrange-tests">Arrange Tests</Link></li>
       <li><Link to="/tests/badge-tests">Badge Tests</Link></li>
       <li><Link to="/tests/base-tests">Base Tests</Link></li>
@@ -24,9 +25,11 @@ const TestLinks = () => (
       <li><Link to="/tests/list-tests">List Tests</Link></li>
       <li><Link to="/tests/loading-overlay-tests">LoadingOverlay Tests</Link></li>
       <li><Link to="/tests/modal-tests">Modal Tests</Link></li>
+      <li><Link to="/tests/modal-manager-tests">Modal Manager Tests</Link></li>
       <li><Link to="/tests/multi-select-list-tests">MultiSelectList Tests</Link></li>
       <li><Link to="/tests/overlay-tests">Overlay Tests</Link></li>
       <li><Link to="/tests/overlay-container-tests">OverlayContainer Tests</Link></li>
+      <li><Link to="/tests/popup-tests">Popup Tests</Link></li>
       <li><Link to="/tests/profile-image-tests">Profile Image Tests</Link></li>
       <li><Link to="/tests/progress-bar-tests">ProgressBar Tests</Link></li>
       <li><Link to="/tests/responsive-element-tests">ResponsiveElement Tests</Link></li>