+++ packages/terra-button-group/CHANGELOG.md
@@ -3,7 +3,11 @@ Changelog
 
 Unreleased
 ----------
+
+1.2.0 - (July 26, 2017)
+------------------
 ### Changed
+* Converted component to use CSS modules
 * Updated spacing around react-docgen comments
 
 1.1.0 - (July 18, 2017)