+++ packages/terra-date-picker/CHANGELOG.md
@@ -3,7 +3,13 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added requestFocus and releaseFocus props to support requesting and releasing focus.
 
+1.2.0 - (July 26, 2017)
+------------------
+### Changed
+* Minor version bump
 
 1.1.0 - (July 18, 2017)
 ------------------