+++ packages/terra-time-input/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+1.2.0 - (July 26, 2017)
+------------------
+### Fixed
+* Fixed onChange callback to occur for onBlur or onKeyDown
+
 1.1.0 - (July 18, 2017)
 ------------------
 ### Changed